saving progress at sph_sim_step() and constrain_vel()

+++ simulation/sph_structs.h
@@ -126,7 +126,7 @@ struct group_conf_struct {
 	// Which group does each loiter circle belong to?
 	// Group number corresponds to array index for num_veh
 	// -1 means not active
-	VectorXd loiter_group;
+	VectorXi loiter_group;
 
 };
